requirements.txt
# requirements.txt
#
# --- Production Dependencies ---
#
# This file lists the absolute minimum packages required to RUN the application
# (e.g., translate.py, suggest.py). It should NOT contain any packages used
# only for development, testing, or analysis.
#
# This is the file that would be used to set up a production environment.
#
# Install with: pip install -r requirements.txt


django>=5.0.0
djangorestframework>=3.15.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.0
google-cloud-storage>=2.16.0
python-dotenv>=1.0.1

python-decouple

django-redis
django-storages

inflect
htmlmin

mutagen